# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  internal_api.createUserRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      role:
        description: '"viewer", "operator", "owner"'
        type: string
    type: object
...
  internal_api.updateUserRequest:
    properties:
      enabled:
        type: boolean
      name:
        type: string
      password:
        description: optional — only set if changing
        type: string
      role:
        type: string
    type: object
  internal_api.userResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      enabled:
        type: boolean
      id:
        type: string
      last_login_at:
        type: string
      name:
        type: string
      role:
        type: string
    type: object
...
host: localhost:6070
info:
  contact: {}
  description: Multi-tenant SaaS platform for satellite device management. Ingests
    MO messages from Iridium/Astrocast, manages devices, SOS escalation, dead man's
    switch, and E2E encryption.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  title: MeshSat Hub API
  version: "1.1"
paths:
...
  /api/users:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_api.userResponse'
            type: array
      summary: List all users
      tags:
      - users
    post:
      consumes:
      - application/json
      parameters:
      - description: User details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api.createUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_api.userResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new user
      tags:
      - users
  /api/users/{id}:
    delete:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete user
      tags:
      - users
    get:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.userResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api.updateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.userResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update user
      tags:
      - users
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"